+++ obs-studio-client/source/nodeobs_service.hpp
@@ -64,8 +64,8 @@ Napi::Value OBS_service_processReplayBufferHotkey(const Napi::CallbackInfo &info
 Napi::Value OBS_service_getLastReplay(const Napi::CallbackInfo &info);
 Napi::Value OBS_service_getLastRecording(const Napi::CallbackInfo &info);
 void OBS_service_splitFile(const Napi::CallbackInfo &info);
-long getServiceIdByName(std::string serviceName);
-std::string getServiceNameById(long serviceId);
+int getServiceIdByName(std::string serviceName);
+std::string getServiceNameById(int serviceId);
 
 Napi::Value OBS_service_createVirtualWebcam(const Napi::CallbackInfo &info);
 Napi::Value OBS_service_removeVirtualWebcam(const Napi::CallbackInfo &info);
